feat: disabling ad ui based on view property
