+++ lib/index.js
@@ -102,6 +102,7 @@ function PlayerView(props) {
   return /* @__PURE__ */ import_react2.default.createElement(NativePlayerView, {
     ref: nativeView,
     style,
+    disableAdUi: props.disableAdUi,
     onDestroy: proxy(props.onDestroy),
     onEvent: proxy(props.onEvent),
     onMuted: proxy(props.onMuted),
