+++ lib/index.mjs
@@ -66,6 +66,7 @@ function PlayerView(props) {
   return /* @__PURE__ */ React.createElement(NativePlayerView, {
     ref: nativeView,
     style,
+    disableAdUi: props.disableAdUi,
     onDestroy: proxy(props.onDestroy),
     onEvent: proxy(props.onEvent),
     onMuted: proxy(props.onMuted),
